refactor: 테스트한 대상의 패키지와 클래스 이름이 일치하도록 리팩토링
