+++ app/src/androidTest/java/woowacourse/movie/presentation/reservation/MovieReservationTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.presentation.reservation
 
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions.click
@@ -9,10 +9,11 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import woowacourse.movie.R
 import woowacourse.movie.presentation.reservation.MovieReservationActivity
 
 @RunWith(AndroidJUnit4::class)
-class TicketCounterTest {
+class MovieReservationTest {
     @get:Rule
     val activityRule = ActivityScenarioRule(MovieReservationActivity::class.java)
